app/map_world/requirements.txt
-i https://pypi.org/simple
asgiref==3.6.0 ; python_version >= '3.7'
branca==0.6.0 ; python_version >= '3.7'
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
django==4.1.7
djangorestframework==3.14.0
environs==9.5.0
et-xmlfile==1.1.0 ; python_version >= '3.6'
folium==0.14.0
idna==3.4 ; python_version >= '3.5'
jinja2==3.1.2 ; python_version >= '3.7'
markupsafe==2.1.2 ; python_version >= '3.7'
marshmallow==3.19.0 ; python_version >= '3.7'
numpy==1.24.2 ; python_version >= '3.8'
openpyxl==3.1.2
packaging==23.0 ; python_version >= '3.7'
pandas==1.5.3
psycopg2-binary==2.9.5
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0 ; python_version >= '3.8'
pytz==2023.3
requests==2.28.2
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.3 ; python_version >= '3.5'
urllib3==1.26.15 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
